tools/model_runner/requirements.txt
firebase-admin==6.5.0
ultralytics==8.3.0
requests==2.32.3
Pillow==10.4.0
python-dotenv==1.0.1
